btrfs-progs: qgroup show: fix formatting of qgroupid on json output
On a 32bit host the split qgroupid is wrong due to the way the numbers
are passed to the formatter as variable length arguments. The level is
u16, promoted to int and then parsed as u64. This means that the values
are shifted and some stack data are printed instead.

Example error messages from yast2-bootloader:

  SystemCmd.cc(addLine):569 Adding Line 7 "      "qgroupid": "21474836480/23885859321282560","

The value 21474836480 = 0x5000000 is 0x5 shifted by 32 bits,
23885859321282560 is 0x54dc1000000000 and shifting by 32 does not
lead to a valid value which should be 0 in this case.

Bugzilla: https://bugzilla.suse.com/show_bug.cgi?id=1209136

+++ cmds/qgroup.c
@@ -1475,7 +1475,7 @@ static void print_all_qgroups_json(struct qgroup_lookup *qgroup_lookup)
 		fmt_print_start_group(&fctx, NULL, JSON_TYPE_MAP);
 
 		fmt_print(&fctx, "qgroupid",
-				btrfs_qgroup_level(qgroup->qgroupid),
+				(int)btrfs_qgroup_level(qgroup->qgroupid),
 				btrfs_qgroup_subvolid(qgroup->qgroupid));
 		fmt_print(&fctx, "referenced", qgroup->info.referenced);
 		if (qgroup->limit.flags & BTRFS_QGROUP_LIMIT_MAX_RFER)
